Optimizing Edmonds-Karp implementation
The Edmonds-Karp implementation has been adjusted so that the flow value
on each edge is not tracked during the main algorithm loop, but to be
calculated in the end, basing on the capacities in the residual network.
This should be slightly faster and seems to produce equivalent results.

Documentation has been adjusted to reflect the changes made.